Begin of the solution to the insertion problem
As a insertion mecanism, it should be implemented in a throught out way,
keeping in mind that, as the organization solution, it will be separated
in many steps, in order to become logic.
So far, I've done an iterator to the references, reverse, one for the
values and is missing, now, a part related to the transition of the
pointers, which can be solved by a highly specialized foreach mecanism,
a more specialized than the current one.