Close to finish the insertion logic
Only missing the final piece of the logic:
The whole insertion works fine as long as it does respect the maximum
capacity per segment, which for the tests is stated as 5, but it will be
changeable.
The logic fails to increase the level count and reallocate the current
reference item to a child of the root.
One thing to be noticed is that the reference node, that will hold the
value, needs to be changed if the level increases. either by getting it
when the mapping occours, either when the level count is enhanced.
Decisions, decisions...